Download TCOON data
This is a function ('Download_Tidal.py') and a call file
('Update_Tidal_data.py') that allow the user to download real-time
surface water data from the Texas Coastal Oceanic Observation Network
(TCOON).

** Note: TCOON had to close due to funding issues in August 2014. An
update is necessary to download the tidal data from NOAA.

The user provides a unique site ID and a time period of data required
(start and end date) as strings. The obtained data is written to a text
file and saved in a specified directory for later, repeated access.

The function is a simple web-crawler application.

The user provides a unique site ID number and a time period of data
required (start and end date) as strings. The obtained data is written
to a text file and saved in a specified directory for later, repeated
access.

The function is a simple web-crawler application.